[construction-set-spf] Edit tests goal constraints
We should specify the locked joints as fully EqualToZero since the
StatesPathFinder is not yet able to initialize/propagate the RHS
properly. A test that SPF is not able to solve is also added.
In the future, defining the goal as set of constraints should
be able to handle more cases like this.